requirements.txt
Django>=4.2,<5.0
gunicorn>=21.0.0
pytesseract>=0.3.10
Pillow>=10.0.0